convert `setDrag` to ts
